% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linktable.R
\name{create_fact_table}
\alias{create_fact_table}
\title{Create a fact table for linktable modelling}
\usage{
create_fact_table(df, key_name, key_columns, drop_columns = NULL)
}
\arguments{
\item{df}{A frame to be transformed.}

\item{key_name}{A string with the new key column's name}

\item{key_columns}{A character vector of columns to be concatenated to generate the new 'key_name' column}

\item{drop_columns}{A character vector of additional columns to drop from the resulting data table. Defaults to NULL.}
}
\value{
A data.table with the new key column and without the original key columns and any additional specified columns.
}
\description{
This function creates a new column in a data frame by concatenating all columns specified in the key list and then drops those columns from the resulting data table. It also allows the user to drop additional columns by specifying them in the \code{drop_columns} argument.
}
\examples{
df <- data.frame(
  uo_cod = c(1501, 1501, 1251, 1251, 1251),
  uo_sigla = c("SEPLAG", "SEPLAG", "PM", "PM", "PM"),
  fonte_cod = c(60, 10, 10, 10, 60),
  vl_receita_exec = c(60, 70, 80, 90, 100)
)

create_fact_table(df, key_name = "chave_rec", key_columns c("uo_cod", "fonte_cod"))
create_fact_table(df, key_name = "chave_rec", key_columns c("uo_cod", "fonte_cod"), drop_columns = "uo_sigla")
}
