% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datapackage.R
\name{read_datapackage}
\alias{read_datapackage}
\title{Criar uma lista de dataframes}
\usage{
read_datapackage(path)
}
\arguments{
\item{path}{Arquivo datapackage}
}
\value{
Uma lista de dataframes
}
\description{
Faz a leitura de recursos de um datapackage e cria um lista de daframes
}
