% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datapackage.R
\name{read_datapackage}
\alias{read_datapackage}
\title{Cria uma lista de dataframes contendo os recursos de um datapackage}
\usage{
read_datapackage(path)
}
\arguments{
\item{path}{Arquivo datapackage}
}
\value{
Uma lista de dataframes
}
\description{
Cria uma lista de dataframes contendo os recursos de um datapackage
}
